WIP: changing structure of instructions
I was planning on using a big switch statement to switch on opcodes and
then just definine the code for each operation inline. This would lead
to a ton of duplicated code though, because of calculations done for
each addressing mode for each instruction. So I am opting for an
instruction struct which holds all the data needed for each instruction,
and then using function calls to implement each op.