Testing all address modes of LDA, success!
This patch begins to organize the cpu and bus to prepare for structuring
a system out of components. The main idea is that the cpu internally
calls bus_read() and bus_write(). These functions are defined by the
bus, which will eventually route reads/writes to the appropriate
peripherals.

This patch also includes a 6502 program for testing all the address
modes of the LDA instruction, called test_lda.o. Loading this program
into memory using the repl, and stepping through it results in the
expected behavior!