Implement more instructions
This patch implements most of the branching, incrementing, and
decrementing families of instructions. It also adds a few small test
programs, but these instructions have not been thoroughly tested yet. I
am testing them a little bit, but mostly expecting to run one of the
well known test suites on the cpu once it is in a complete enough state.