Cartridge module
This patch adds a cartridge module that can load program data from an
an ines file. Currently the implementation is incomplete, I just wrote
enough to allow me to load the program data into memory so that I can
execute it on the emulated 6502. I am mostly using this to run
nestest.nes to test my cpu.