Illegal Opcodes
This patch implements all of the illegal opcodes.

There are also a few bug fixes in some of the instructions.

This patch also introduces debug logging of each instruction. The format
is the same as nestest.nes since this is the test I am testing against
currently.

nestest finally seems to be passing with no failures as well!