Remove repl, preparing for graphics
This patch is preparing for doing graphics.

We were working with the whole address space as just a flat array of
memory. Now were allowing the bus to route addresses to different
peripherals.

The cart has been updating to prepare for implementing different
mappers, right now it is just a temp hack to get things working.