Implement mapper 0
This patch introduces mappers. Reads and writes of the cartridge as well
as on the internal ppu bus have now changed to be under mapper control.

Now that horizontal and vertical nametable mirroring is controlled by
mappers, I am seeing a one pixel line on the left of the screen (for
example in donkey kong) that seems to be incorrectly reading from
nametables. If I switch donkey kong to vertical mirring mode (which is
not the correct mode for dk) it looks fine so I don't know if it is just
horizontal mirroring that is broken or what.