Fix sprite evaluation off by 1 pixel
I don't understand why the current scanline is used. All documentation I
can find on the NES says that during a particular scanline, the ppu
calculates the sprites that will be visible on the next scanline.
However, when testing whether the sprite appears on the NEXT scanline,
the sprites get rendered 1 pixel to high. If the current scanline is
used then it renders correctly. Does this mean that all NES games
actually keep track of y position 1 pixel higher than their actual
position to compensate for this??? Or am I just stupid and missing
something else?